SimDRAM: Use svGetArrElemPtr1 instead of svGetBitArrElemVecVal
Xcelium doesn't support `svGetBitArrElemVecVal`, and segfaults when that
function is executed. Using `svGetArrElemPtr1` should have the same
behavior. It returns a pointer to the array element, rather than
`svGetBitArrElemVecVal`, which returns a copy of the array element. Also
`svGetBitArrElemVecVal` works with a wider variety of bit-array data
types, whereas `svGetArrElemPtr1` seems more useful when the underlying
datatype can be represented as a native C datatype. Luckily, the
underlying SystemVerilog type here is `byte`.